[feat] 스와이프 예시 구현
