+++ client/package.json
@@ -26,6 +26,8 @@
     "react-native-responsive-screen": "^1.4.2",
     "react-native-safe-area-context": "^3.2.0",
     "react-native-screens": "^3.4.0",
+    "react-native-swipe-gestures": "^1.0.5",
+    "react-native-swiper": "^1.6.0",
     "react-native-vector-icons": "^8.1.0"
   },
   "devDependencies": {
